Refactor: rename `DataTable` to `Datatable` and standardize related imports ♻️
- Renamed `DataTable` component and type definitions to `Datatable` for consistency.
- Updated all associated imports and file paths to follow a uniform lowercase naming convention.
- Adjusted stubs, utils, and regex patterns to reflect the changes.
